make fromResult more flexible
It was too hard to compose except in the specific case where could accept a
literal string and do something with it. If you needed any pre-processing
whatsoever, you had to drop back to an `andThen`. This makes that step a little
easier, even though if you have something like `Hex.fromString` you need a
little more ceremony.